+++ src/vector.c
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-Vector *VectorNew(int elem_size)
+Vector *VectorNew(size_t elem_size)
 {
     Vector *v = (Vector *)malloc(sizeof(Vector));
     v->elem_size = elem_size;
@@ -32,12 +32,12 @@ static void VectorExpand(Vector *v)
     }
 }
 
-int VectorLength(Vector *v)
+size_t VectorLength(Vector *v)
 {
     return v->logicl_length;
 }
 
-void *VectorNth(Vector *v, int index)
+void *VectorNth(Vector *v, size_t index)
 {
     return (char *)v->elems + v->elem_size * index;
 }
@@ -54,9 +54,9 @@ void VectorAppend(Vector *v, const void *value_addr)
 
 void VectorMap(Vector *v, VectorMapFunction map, void *aux_data)
 {
-    int length = VectorLength(v);
+    size_t length = VectorLength(v);
 
-    for (int i = 0; i < length; i++)
+    for (size_t i = 0; i < length; i++)
     {
         void *value_addr = VectorNth(v, i);
         map(value_addr, i, v, aux_data);
@@ -66,11 +66,11 @@ void VectorMap(Vector *v, VectorMapFunction map, void *aux_data)
 Vector *VectorCopy(Vector *v, VectorCopyFunction copy_fn, void *aux_data)
 {
     Vector *new_vector = VectorNew(v->elem_size);
-    int length = VectorLength(v);
+    size_t length = VectorLength(v);
 
     if (copy_fn != NULL)
     {
-        for (int i = 0; i < length; i++)
+        for (size_t i = 0; i < length; i++)
         {
             void *value_addr = VectorNth(v, i);
             AST_Node *binding = *(AST_Node **)value_addr;
@@ -82,7 +82,7 @@ Vector *VectorCopy(Vector *v, VectorCopyFunction copy_fn, void *aux_data)
     else if (copy_fn == NULL)
     {
         memcpy(new_vector, v, sizeof(Vector));
-        int elems_size = v->allocated_length * v->elem_size;
+        size_t elems_size = v->allocated_length * v->elem_size;
         void *elems = malloc(elems_size);
         memcpy(elems, v->elems, elems_size);
         new_vector->elems = elems;
@@ -97,8 +97,8 @@ int VectorFree(Vector *v, VectorFreeFunction free_fn, void *aux_data)
     
     if (free_fn != NULL)
     {
-        int length = VectorLength(v);
-        for (int i = 0; i < length; i++)
+        size_t length = VectorLength(v);
+        for (size_t i = 0; i < length; i++)
         {
             void *value_addr = VectorNth(v, i);
             free_fn(value_addr, i, v, aux_data);
