+++ src/vector.h
@@ -1,18 +1,20 @@
 #ifndef VECTOR
 #define VECTOR
 
+#include <stddef.h>
+
 typedef struct _z_vector {
     void *elems; // created by malloc, needs to be released.
-    int elem_size;
-    int logicl_length;
-    int allocated_length;
+    size_t elem_size;
+    size_t logicl_length;
+    size_t allocated_length;
 } Vector;
-typedef void (*VectorFreeFunction)(void *value_addr, int index, Vector *vector, void *aux_data);
-typedef void (*VectorMapFunction)(void *value_addr, int index, Vector *vector, void *aux_data);
-typedef void *(*VectorCopyFunction)(void *value_addr, int index, Vector *original_vector, Vector *new_vector, void *aux_data);
-Vector *VectorNew(int elem_size);
-int VectorLength(Vector *v);
-void *VectorNth(Vector *v, int index);
+typedef void (*VectorFreeFunction)(void *value_addr, size_t index, Vector *vector, void *aux_data);
+typedef void (*VectorMapFunction)(void *value_addr, size_t index, Vector *vector, void *aux_data);
+typedef void *(*VectorCopyFunction)(void *value_addr, size_t index, Vector *original_vector, Vector *new_vector, void *aux_data);
+Vector *VectorNew(size_t elem_size);
+size_t VectorLength(Vector *v);
+void *VectorNth(Vector *v, size_t index);
 void VectorAppend(Vector *v, const void *value_addr);
 Vector *VectorCopy(Vector *v, VectorCopyFunction copy_fn, void *aux_data);
 void VectorMap(Vector *v, VectorMapFunction map, void *aux_data);
